refactor and clan up evaluator export functions
